Implemented a purely virtual VQwBCM class so that the QwBeamLine can work with either VQWK or Scaler channels seemlessly.
QwBeamLine now stores the BCM's as a vector ofshared_ptr<VQwBCM*>,
that can be either VQWK or one of the many Scaler channels.
Similarly for QwCombinedBCM.

To add additional scalers one should add to the Create factory
function the appropriate scaler (or other templated type) channel.

Has been tested to produce ROOT files that are consistent with
revision 2957 of this branch.

Main achievement: QwBeamLine BCM's work for the Compton!

git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/pking.templated_beamline@2958 b9700a98-9937-0410-b777-f318d6eb5bff