Additional modifications of qwbatchsub.pl to resolve some error messages from the use of undefined variables.
If the ROOT file output path is not defined, it will default
to the pass0 MSS directory; this is not the best default, but
it will do for now.
People doing there own analysis should try to use the work disk
if they can.

Removed some relic code from the Gzero version of batchsub; this
was related to passing extra environment variables to the batch
job.  We don't use those variables anyway.

Increased the diskspace requested to be closer to our actual usage;
this shouldn't cause problems, but I had one job die when it ran
out of local diskspace.
This should make sure we're running on machines that can actually
hold our ROOT files.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2966 b9700a98-9937-0410-b777-f318d6eb5bff