Fixed compilation warning in QwEventBuffer.
Added option -a or --add-config that attempts to resolve some of the
ambiguity in the -c or --config flag.  The option add-config will add the
specified config file WITHOUT throwing out the default config files.  This
can be used to add bits of configuration stored in .conf files.  Also made
the usage info a little bit clearer for both cases.
Note: This has no effects on how --config is processed.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2983 b9700a98-9937-0410-b777-f318d6eb5bff