More changes to the way magnetic fields are handled.  This can now easily read the more recent field maps that were obtained from the JLab mapping measurements (but unfortunately are in a different reference frame, different grid, and at a different current than the older one...).
As already stated in a previous commit, there is one config file for each field
map.  To use fieldmap peiqing_2010_zc.dat, use --add-config peiqing_2010_zc.conf.
This will make sure that the particulars of that map are taken into account.

Because BFIL becomes quite meaningless in the presence of these different maps
("BFIL with respect to which map?"), you should now specify the current.  The
config files above contain the current at which the map was taken, and the BFIL
is determined from that.  The default current is set to 8920 A, so normally you
don't need to specify this at all.

The new syntax is now:
  bin/qwtracking --add-config peiqing_2010_zc.conf --QwMagneticField.current 8765 ...

The old syntax still works (except for QwMagneticField.bfil introduced yesterday),
e.g. bin/qwtracking --QwTracking.fieldmap peiqing_2007.dat ...

Since reading the magnetic field map takes a while, the code automatically tries to
find a binary fieldmap with extension .bin in the same place as the specified map.
If it is found, the analyzer will read that map.  If not, it will just read the
usual .dat file.  You can generate your own binary field map using the conversion
utility QwMagneticFieldConvert:
  bin/qwmagneticfieldconvert --add-config peiqing_2010_zc.conf --convert

Note: if you use either of the newer 2010 fieldmaps, you will get a lot of output
that indicates that there are (intentional) holes in the map.  Just ignore that.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2985 b9700a98-9937-0410-b777-f318d6eb5bff