an initial commit of the hall A regression code.
There are several issues that need to be worked on:
The segment number is hard coded in TaInput.cc as 000 -- this means that if you want to look at a different segment you will 
have to recompile.

The directories in the run_redana.C macro need to be updated for cdaq use.

The control.conf is minimal it can be expanded.  It should be updated to agree with Jan's code.

The customcut in the control.conf is minimal and can be expanded.  This should be updated to agree with Jan's code.

The event number 'ev' is not filled properly.

Other things I haven't noticed yet.  I just started looking into this code...


The pan directory is needed for now.  You must make that dir then make the postpan dir.




git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2995 b9700a98-9937-0410-b777-f318d6eb5bff