Note: this is branch pking.templated_beamline.
Implemented a proper time normalization for the Scaler
channels using a normalization from a new beamline class
VQwClock (and a templated subclass QwClock). For now,
only implemented on BCM scalers, but could easily be
changed for all other scalers.

Scaler clocks could be added as detector type clock in
the channels map file. Setting the pedestal to zero and
the gain to the standard frequency of the clock. A new
valued pair clocknorm=clock_name, where clock_name is
the clock to be normalized with. As an example one may
be the compton_beamline channel and maps.

Changed the default scaler to be of type SIS3801D24 in
the BCM's and BPM's.

Added new functions to VQwDataElement to allow a clock
to be assigned to any scaler.

The output of a VQwClock are in units of time (s).
The output of an uncalibrated BCM are now in units
of Hz (instead of counts). Keep this in mind when
performing a BCM calibration, one no longer needs
to perform a clock calibration at the same time.

Minor output cleanup.  Made some functions more silent.

git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/pking.templated_beamline@3003 b9700a98-9937-0410-b777-f318d6eb5bff