Geometries for region 2 are now read from the qweak_geo.new file. This is the same file that QwTracking uses to read in the geometry. By looking through tracking run 8658 I see "clean" partial tracks through Region 2.
Region 3 geometries are still hard coded in, and presently not
up to date with the true geometries. This will change once I
read from the same file as Region 2.

I removed the Region 2 frame for now, needs some caring to be
geometrically correct again.

Also updated is the GDML file. At this point is is very different
than the GEANT4 output. But now has two sets of HDC's and VDC's.

git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3004 b9700a98-9937-0410-b777-f318d6eb5bff