Removed three dependency files from HallA/pan/codaclass; these are regenerated during compilation.
Modified all of the PAN Makefiles to use "root-config --incdir"
to get the include file path instead of relying on it being
"$ROOTSYS/include" (which is not correct for the CUE installations).

Added a Makefile to the base pan/postpan directory to only call
the builds for the sections of pan/postpan needed to create
the linear regression analysis tool.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3010 b9700a98-9937-0410-b777-f318d6eb5bff