Changed the default database displayed by the GUI to be the qw_run1_pass3 DB on qweakdb.
Added some functionality to select the good_for and run_quality
in the GUI; it is not yet modifiable, but the structure is more
capable of allowing it to be modified by the user.  Force the
good_for to be non-NULL and to include either (1) or (3).
Force run_quality to be non-NULL and include (1).

Added a basic test on Wien reversal; for now the sign of the
Phi_FG gives the Wien reversal.  This is now part of the
plotting routines.

Added functionality to plot regressed variables and a selection
box for "correction" or "no correction".

Started to add functionality to plot histograms of the distributions
of the variables for each runlet.  This is not complete, and is just
a copy of the plotting routine versus run number right now.

Split out the DB query from the plotting routine to allow the
histogram and time-ordered plotting routines to use the same
query routines.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3016 b9700a98-9937-0410-b777-f318d6eb5bff