Modifications to linRegBlue to run the Pass3 analysis.
The routine to determine the initial ROOT file name in order to
create a regressed ROOT file now will find the "basename" of the
initial ROOT file name instead of looking for "/Qw".  In the
Pass3 analysis environment, there are directories starting with
"/Qw".

Changed the independent variable definitions to use the new target
variables.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/balewski.correlator@3021 b9700a98-9937-0410-b777-f318d6eb5bff