Added an automatic job-segmentation scheme to qwbatchsub.pl, which will generate independent jobs for subsets of segments for a run which would require more disk space than the batch system maximum permitted.
Added a "segments" option to qwbatchsub.pl.  It will
take either single segements, a comma-seperated list,
or a colon-seperated range of segments.
If you specify the segments, you must do it for a 
single run.

In either of these cases, the XML job file and the 
log files would be created with the segment range
in their name.  The job name also contains the
segment range.


This has not yet been fully tested.
Only the creation of the XML job files has been tested;
the generated job files have not been submitted to the
batch system.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3074 b9700a98-9937-0410-b777-f318d6eb5bff