Modifed the doubleRegPass_farm.sh script.
Old directories will be renamed to contain a timestamp of their
last modification time.

We don't try to change the permissions on the old directories and
their contents, because if there is a problem with the permissions
we won't be able to change them anyway.

All new directories created in the script will also contain a
timestamp of their creation time.

New directories and their contents will be forced to have "c-qweak"
group ownership, and to be owner and group writable.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/balewski.correlator@3094 b9700a98-9937-0410-b777-f318d6eb5bff