A first attempt at including the run_quality cuts into QwDBGUI. It appears to be working, but it's far from impossible that there are some special cases that it doesn't handle properly. For instance, I've yet to nail down exactly what should be displayed if a run has run_quality_id=={1,3} meaning that it was flagged as both GOOD and SUSPECT. My gut is telling me that it should be displayed as SUSPECT, and that likewise any set containing run_quality_id==2 should be displayed as BAD. I will continue to work on making that so (if it doesn't already happen to act that way).
This commit also includes changes that add the capability to plot the 7 double difference channels we have. They are:

Main Detectors
	dd_mdeven_odd
	dd_md15_37
	dd_md26-48
BCMs
	bcmdd12
	bcmdd15
	bcmdd16
	bcmdd56

Where the Main Detector DD's are in the database with "Measurement Type" == "a" and the BCM DD's have "Measurement Type" == "d".

For now the users will just need to "remember" that until I can find a way to have the GUI pick up on the fact that a DD channel has been picked and force the correct Measurement Type.

PLEASE NOTE:
We've recently discovered that it looks like the double differences did NOT get written to the database during the Pass3 Replay, so if you try to plot these values now it will tell you that there is no data on the database.

git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3108 b9700a98-9937-0410-b777-f318d6eb5bff