NOTE:  This is on branch wdconinc.first_pass_regression.
Updating the branch to contain the full change set associated
with branch pking.gcc_346.
This was done in a round-about process to resolve errors:
1)  Start with branches/wdconinc.first_pass_regression@3062
2)  Merge in changes from revisions 3060:3081 on the trunk
3)  Merge in the differences between trunk@3081 and branches/pking.gcc_346@3109
4)  Update to branches/wdconinc.first_pass_regression@3109

In the process of the last update, I discovered that the
QwSubsystemFactory.h committed by 3109 was an accidental
overwrite with the contents of QwSubsystemArray.h.
The QwSubsystemFactory.h has been restored to the comparable
version as in branches/pking.gcc_346.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/wdconinc.first_pass_regression@3111 b9700a98-9937-0410-b777-f318d6eb5bff