Some adjustments to how the run_quality_id was being handled by the GUI. Like I mentioned in my previous commit, I was unsure how it was dealing with entries that users described as being "good AND suspect", "suspect AND bad", and even (gasp) "good AND bad". I've made it explicitly choose the following classification scheme:
Anything labeled as bad (bad, good+bad, suspect+bad) will be considered to be bad by the GUI.
Anything labeled as suspect but NOT bad will be considered suspect.
Only events with run_quality_id==1 (not 1,2 or 1,3) are considered good.

My previous version ignored any entries that had more than one run_quality_id: so this commit obviously fixes that too.

This commit also includes handling of the good_for_id fields. I added in a TGListBox with our different good_for types and put it into the GUI in such a way that I don't feel clutters it up too much. Multiple good_for types may be selected simply by clicking on them. The default is to have PRODUCTION and PARITY selected.

Similarly to the run_quality_id problem: runs here often have more than one good_for_id associated with them (such as {1,3} for production parity, or {1,4} for production tracking). This is dealt with in what I considered to be the obvious fashion, using queries like the following snippet:

...
WHERE good_for_id NOT LIKE "%2%" AND good_for_id NOT LIKE "%5%" AND good_for_id NOT LIKE "%6%" AND good_for_id NOT LIKE "%7%" AND good_for_id NOT LIKE "%8%"
...

This means that any run with good_for_id containing a 1,3,or 4 will be plotted. (1, 3, 4, {1,3}, {1,4}, {3,4}).



For people who want to give it a test run, the run range 11400-11450  contains "good parity" "suspect parity" and "suspect tracking" runs.

git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3112 b9700a98-9937-0410-b777-f318d6eb5bff