Changes to the QwDatabase GUI
This version implements the ability to double click on the plot in QwDatabaseGUIView
which opens a separate window with the current data. This data window will not be 
affected by subsequent queries and allows the user to save the data from the plot in 
row column format.

More changes will come soon ...  



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3117 b9700a98-9937-0410-b777-f318d6eb5bff