This commit introduces plotting Main Detector and LUMI sensitivities from the database.
I have added two new INSTRUMENT TYPEs called "Main Detector Sensitivities" and "LUMI Detector Sensitivities". Upon selecting one of these, the "Detector Property/Regression IV" combobox will be enabled and filled with our 5 regression iv's (targetX, targetY, targetXSlope, targetYSlope, energy). The MEASUREMENT TYPE is set to "s". This was required to properly label the y-axis.

I've yet to perform a complete set of tests (for instance, looking at lumi sens wrt all variables for different goodfor and runquality settings), but it appears to be working fine for most of the standard queries.

-------

This commit also includes a bit of clean-up that was causing the GUI to segfault if the number of IHWP configurations to be displayed changed between queries.

-------

A big thank you to Buddhini for giving me a hand with these updates!

git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3121 b9700a98-9937-0410-b777-f318d6eb5bff