A first pass consistency check for the older (pre run 9000) main detector pedestal files. I think I picked up all of the errors which included:
A number of runs fit into the run ranges for two pedestals files. The analyzer is smart enough to handles these in a smart way (it chooses the pedestal file with the smallest valid run range for a given run), but we can't get into trouble if we just keep the ranges clean in the first place.

A number of the oldest pedestal files had channel names appropriate for an older version of our channel map, either not having the qwk_ prefix or not having the "md" portion.

I also commit a macro that will loop through all main detector pedestal files and change the pedestals to a user-given number. This was used to set the current sources' pedestals to zero, but I can imagine it being useful again in the future so here it is.

Cheers

git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3127 b9700a98-9937-0410-b777-f318d6eb5bff