Merged the changes to qwbatchsub.pl from trunk revision range 3078:3140 into branch release_1.09.
This gives us the capability of only submitting jobs for
runs not marked "bad" in the database.

Although, if we were overzealous at marking runs "bad", we
may not want to use this capability just yet (and a quick 
scan through the DB suggests that is the case right now).


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/release_1.09@3141 b9700a98-9937-0410-b777-f318d6eb5bff