Created a script to determine the amount of transverse polarization of the beam based on the regressed main detector asymmetries. Currently, the script is accessing the run1 pass3 database to get the regressed pass3 data. The idea is to perform a cosine fit on the main detector bar asymmetries over the azimuthal angle using the measured transverse asymmetry from the transverse measurement for LH2 cell and take the amplitude of that fit as the aprecentage of residual transverse polarization in the beam. I testesd with the transverse data and ofcourse got ~99% polarization.
Fit = (A_T_transverse)*p0* cos(phi+p1) + p2
when A_T_longitudinal = A_T_Transverse*p0 with p0 giving the amount of transverse polarization.

Testing is underway using the longitudinal data before the transverse running where we know there was a 2% transverse polarizaiton in the beam.
But with IN+OUT being !=0 for taht data set, it is hard to get the results I expected to get because my script assume that the only false asymmetry in the data that has an azimuthal variation is the transverse asymmetry.
 


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3162 b9700a98-9937-0410-b777-f318d6eb5bff