NOTE:  this is in branch wdconinc.first_pass_regression.
This is a partially complete single-event regression framework.
A "regression susbsystem" is created, and appears to be getting
instantiated correctly, but the variables are not yet being passed
into the ROOT trees, so detailed testing has not been completed.

This is the last commit from S. Meeker's work over the summer.
The engine compiles and runs; however, there is a segfault
during the cleanup that we did not have time to address.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/wdconinc.first_pass_regression@3175 b9700a98-9937-0410-b777-f318d6eb5bff