Added a new boolean option "helicity.toggle-mode" to the QwHelicity class. This basically combines a pattern-size of two with a helcity delay of zero, and disables the predictor, to directly use the reported helicity as the "true" helicity.
This should make it simpler to analyze toggle-mode data.

Note that this disables some of the data robustness checking,
as it does not expect toggled pairs to always start with
the same helcity state.  That could be developed more fully
later.

It does verify that each pattern has equal numbers of
positive and negative helicity states though.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3177 b9700a98-9937-0410-b777-f318d6eb5bff