Fixed the fitting limits of the phase in the transverse monitor plot. Now, the script first carries out the transverse fit of the IHWP IN values and then use the fit results from that to initialize, p0, p1,and p2 of the fit of IHWP OUT values. Also, the fit limits of p1 in IHWP out fit it now set to p1_out_lower = p1_in - 90, p1_out_upper = p1_in + 90. This range is chosen because we don't expect the phase of the transverse to change from IN to OUT but in anycase, it is possible that it can change by +- 90 so we'll go from, e.g. being verticle to horizontal or vise versa.
Added two new scripts to monnitor the phase change in transverse polarization during a range of runs and during a range of slugs. 
This will allow us to monitor the phase changes and decide when we apply corrections, do we want to apply them to a slug or a run. 
What ever time scale we choose depends on the stability of the phase of the transverse polarization within that time.  

More information on how to use this scripts can be found inside the scripts.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3186 b9700a98-9937-0410-b777-f318d6eb5bff