Added copying of histogram pointers to operator= for QwHelicity. This allows for filling histograms after popping an event out of the event ring instead of before.  Josh, let me know if this fixes your problem.
On operator=.  We are VERY sloppy everwhere in the code when it 
comes to assignment operators (operator=) and copy constructors 
(QwSomething::QwSomething(const QwSomething&)).  When we do have 
assignment operators, they often do not call the parent class. We 
have something that looks like a copy constructor but takes a 
pointer (and therefore QwSomething a(b) will NOT use that 
constructor).  With the event ring push() will call the assignment 
operator but, depending on how it is called (by ref or by value), 
pop() will use also the copy constructor.  We have private 
assignment operators (which is a hack to force the implementation of 
the assignment operator of derived classes, where you then 
explicitly neglect to call back the parent class).  We have 
assignment operators in the parent of that class which are 
impossible to get to...



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3189 b9700a98-9937-0410-b777-f318d6eb5bff