Ok, here's an *ugly hack* to keep Josh's development work going...
And with this I've got Paul's attention :-)

What is the goal?  Use the beam trip/stability event cuts to prevent
MPS histograms from being filled when the beam is bad!  This is done
by moving the FillHistograms calls after the event ring pop calls.

What are the problems?  Push and pop use a mixture of assignment
operators and copy constructors to get the event in and out of the
ring.  These operations do no work as expected in that they do not
copy the histogram pointers.  Effectively you lose access to the
histograms when pushing and popping.

Why was this not a problem before?  Without push and pop we are
filling the histograms on the exact same object that we created the
histograms with.  The pointers to the histograms are valid for that
objects.  Only when pushing and popping do these pointers become
invalid.  The histograms that *are* being filled in the push/pop
section are stored separately in the QwHelicityPattern object.

So, write valid push/pop methods, then!  Turns out that the trouble
starts in QwVQWK_Channel::operator= (let's ignore the copy constructor
for now, we can call pop by reference and avoid that can of worms).
QwVQWK_Channel::operator= does not call the parent operator= method.
It actually cannot call the parent operator= method because it is
private and not implemented (Paul, was this intentional? This forces
the implementation of the data channel operator= methods because the
default implementation will attempt to cascade down.)  The grandparent
operator= is implemented but also private (and therefore inaccessible).
I removed the operator= methods in the parent classes (so the compiler
will generate suitable ones), and call them from QwVQWK_Channel
Inheritance:
  QwVQWK_Channel <-- VQwHardwareChannel <-- VQwDataElement

But who owns (semantically) the histograms?  The TH1 pointers are stored
in VQwDataElement, so would seem to be owned by that class.  They are
(now) copied on assignment, but should (therefore) not be deleted in
the destructor.  The DeleteHistograms method should take care of that
(but that can be removed from all derived classes and be a regular
method of VQwDataElement).  DeleteHistograms should be called
explicitly on the original object for which the histograms were
created (as we always did).

Does this solve the problem of push/pop?  No, there is creative copying
and assigning going in the BPM stripline class.  That unfortunately
leads to lost histogram pointers and multiple histogram pointers when
VQwDataElement owns the TH1 pointer vector.  It goes something like this:
  VQwDataElement tmp;
  tmp = fWire[0]; // histogram pointers in fWire[0] copied to tmp
  fRelPos[0] = tmp;
  fRelPos[1] = tmp;
  // both fRelPos end up with the histogram pointers of fWire[0]
Because of this I had to disable the DeleteHistogram calls for these
stripline classes in QwBeamLine (temporarily!).

Naturally this still segfaults somewhere else as well, but only at the
very end when main() ends and the event ring destructor is called.
I am going to continue looking into this to find a better solution.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3190 b9700a98-9937-0410-b777-f318d6eb5bff