Modification to the way histograms are handled.  This removes the need to implement or call DeleteHistograms (unless, of course, you create a histogram that isn't written to a ROOT file).  ROOT owns the histogram pointers, and is responsible for cleaning them up when the file is closes.  We just keep a vector of the histogram pointers (trough MQwHistograms).  Both VQwDataElement and VQwSubsystem have histogram functionality.
The only place where some confusion could occur is in the BPM handling.
In order to not overwrite the histogram pointers (and some other values,
such as event cuts) you should not write:
  position = wire1 - wire2;
because the question is then whether the histograms from wire1 or wire2
will be copied into those of position.

This commit was verified on the running sums of all channels, by checking
the histograms and the trees, etc.  There is still a 250 ppm difference in
the filesize of the resulting ROOT files and searching for the cause of
that makes you appreciate precision measurements even more.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3205 b9700a98-9937-0410-b777-f318d6eb5bff