Modified the QwBlinder so that the name of the object is not used by the ReturnInternalValue call when requesting the beam charge from the subsystem array.
This resolves the following error messages produced by the blinder:
	Warning: QwSubsystemArray::ReturnInternalValue: name "qwk_charge" 
	not found in array.
	VQwDataElement::ReturnInternalValue for value name, qwk_charge 
	define the routine in the respective subsystem to process this!


NOTE:  The histogram filling routines still have a problem, which is
not resolved by this commit.

The histogram pointers are reassigned across the "operator=" methods,
and this results in the HelicityPattern's yield and asymmetry objects
filling their data into the single MPS subsystem array's histograms.

I (Paul) think that the mathematic assignment used in calculating
the asymmetries should be handled by a new "AssignValue" method that
will need to propagate through all subsystems and datamembers.
We should examine where the "operator=" method is used, and what
is its utility; it may be that after creating and using an "AssignValue"
method, the "operator=" is only used as a type of "Copy", and could
be entirely replaced by that method.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3219 b9700a98-9937-0410-b777-f318d6eb5bff