Cleaning up the uses of "Copy" and "operator=" in the hardware channel classes.
"Copy" will make an exact copy of the object, and
pointers will be copied.

"operator=" is the arithmetic assignment function,
only the data memeber which are event-based should
be copied; any data memebers which represent the
object configuration are left alone.

This distinction ought to be pushed out through the
subsystem classes, and the derived data elements;
however, with just these changes, the histograms 
and trees produced by this version are consistent
with the histograms and trees produced by r3188.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3224 b9700a98-9937-0410-b777-f318d6eb5bff