Found and corrected an inconsistency in how offsets are applied to VQWK channels.
What had been done was that the offsets applied to the 
subblocks was "offset/fBlocksPerEvent"; but since r1870
the subblock values are normalized to contain the same
value as the HW sum, and so the same offset should be
applied.

This change almost only effects the absolute BPM positions
reported.  The inconsistency led to the average of the
subblocks for the absolute BPM positions being not the
same as the HW sum absolute BPM position.

The changes are in the subblock yields of the BPM positions,
and in the subblock results for the combined BPM and energy
variables.  The subblock "diff" for the energy variable is
changed, but no other diff values are changed by this correction.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3236 b9700a98-9937-0410-b777-f318d6eb5bff