Enabled plotting of weighted average value of a variable within a run (runlet average) vs the time the run started.
The run start time in DB is the run start time recoreded by CODA and currently we only have the run start and run end times and no record of the individual start and end times of the runlets.
When they become available in the next run I will revisit this to enable plotting of runlet value vs runlet time.

To do the run value vs run start time  I used the time display mechanism in TAxis class but unfortunately there seems to be a bug in this class which is correcting the given times for the day light saving. 
There were two ways to fixt this:
1. use text labels on the axis
2. Hardcode the DST start and end times for 2010-2012 and checking the values in the time array and adjusting them +- 1 hour.

I chose to do the 2nd because to me was much easier to do than rearanging the whole function which is used to plot Y vs X for different X (run number and slug number). 
I checked this for runs taken before DST and after DST and to me it looked to be working fine.

Some minor adjustments to the plots and labels were also done to make things look nicer.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3252 b9700a98-9937-0410-b777-f318d6eb5bff