Updated linRegBlue to compile on gcc 4.4.5.  The GCC team decided to become more pick about the matching constancy of a returned char.
Also,
Increased the max number of events in  doubleRegPass.sh.  The number was reduced to 22233 in version 3114. This caused only 20% of the root file to be regressed (22233 of 94766).

The number is now 2223344 to match doubleRegPass_farm.sh.

No, I do not know why this number was chosen.  Maybe Jan is a big fan of Capital City Vacuums.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/balewski.correlator@3275 b9700a98-9937-0410-b777-f318d6eb5bff