Configuration file for the 2011 field map.  Use commands such as   bin/qwtracking -r 8658 -e :1000 --add-config peiqing_2011.conf \     --QwTracking.disable-momentum false --QwTracking.disable-matrixlookup
You can convert this field map in the faster binary format with the command
  bin/qwmagneticfieldconvert --add-config peiqing_2011.conf --convert
This will automatically place the binary file where it belongs, in $QW_FIELDMAP,
and where it will be found by the analyzer.

This configuration file (and its sisters peiqing_2007.conf, peiqing_2010_zc.conf,
and peiqing_2010_ff.conf) work just as well with original files and binary files.
The binary files are just faster and contain the reference current and grid size.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3302 b9700a98-9937-0410-b777-f318d6eb5bff