This commit mainly introduces the 100k auto update for the GUI:
After implementing the auto update in several ways and trying
it on my computer, I have settled on the following scheme:

There is a new (small) executable called "QwAutoGUI" which
serves as a wrapper for the full GUI and does the following:

     1) When started, it checks if there is already an auto
        update process running and asks if that one should be
        replaced with a new one. This makes sure
        that there is never more than one instance of this
	process running.

     2) The program starts the actual GUI as a child process, 
        with the flag "-u true" letting the GUI know that it 
        is to expect signals when a new run is found in the 
        QW_ROOTFILES directory. 

     3) The porgram has a while loop that checks QW_ROOTFILES for
        new files at fixed intervals (currently set arbitrarily to 
	60 seconds ... but can certainly make that longer). If 
	QwAutoGUI finds a new file (with a higher run number) then
	it sends a signal to QwGUI to open the file and process it.

When either QwGUI or QwAutoGUI are killed, they will kill the other
as well.

I have decided against having the QwGUI query the QW_ROOTFILES 
directory directly, because this created interference in the user 
interacting with the GUI and processing the plots, etc ...

I believe this way of doing it will create little overhead and be
convinient.

QwGUI can still be run in the familar way, without any commandline
flags.

------------------------------------------------------------------

I have removed the QwDatabase subsystem from the QwGUI, since we
have now separated the two ...





git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3321 b9700a98-9937-0410-b777-f318d6eb5bff