updated pedestal extraction macro to conform to tighter restrictions on string and character manipulation.
added pedestal difference macro based on the pedestal script.

to calculate the pedestal differences the following works with the trunk:

qwparity -r11503 -e:5000 --rootfile-stem diff_ --enable-differences 1 --disable-histos 1

There is a problem with the histogram handling with differences enabled so disable them or get a seg fault.

The md and bg pedestal differences leaves are empty...  Paul said they may be zeroed out in the analyzer to hide the blinding factor.





git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3355 b9700a98-9937-0410-b777-f318d6eb5bff