Updated and cleaned up the plotting functions to make them be more independent of each other.
Created a new function to get data selections from the comboboxes and NumberEntries and fill the values to a set of global variables that represent the measurement_type, device, device_id etc which can then be used by all the functions that plot data. Previously each function were grabbing these data from the comboboxes/NumberSelection boxes which made the functions sort of depend on the readings of the comboboxes/NumberEntries. With this modification, the functions are now independent of what is set in input widgets. They only rely on the set of global variables to get what they want and these variables can be reset anyway we want. 

Hopefully this will make it easy to add new changes to the GUI.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3373 b9700a98-9937-0410-b777-f318d6eb5bff