This Macro create plots for beamline optics using beam modulation data from Mps_Tree.
To run the script, open root in a terminal in any cdaq machine and execute the macros.
It will ask for a 'run #'. Type 'run #' and press 'ENTER'. Plots, run information and fit parameters will be saved acording to run # in the directory: '/u/group/hallc/www/hallcweb/html/qweak/bmod/images/run#'. The script bmod_beamline.C will use the text file and create necessary files and submit it to BMod website: https://hallcweb.jlab.org/qweak/bmod/

This is the priliminary version of the script. It does not run automatically now. We will add that functionality latter.

--Nuruzzaman 



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3378 b9700a98-9937-0410-b777-f318d6eb5bff