Created a Summary Tab for the QwRealTime GUI using the class QwGUISummary This Summary tab contains (currently) 5 plots that we think are the important ones to check in realtime. These are: 1. asym_qwk_charge_hw 2. diff_qwk_targetX_hw 3. diff_qwk_targetY_hw 4. diff_qwk_bpm3c12X_hw 5. asym_qwk_mdallbars_hw
Next I plan to add a bar chart displaying the number of events in each of these channels with non-zero ErrorFlag and Device_Error_Code. This will be the error summary plot.

The QwGUISummary class uses a single plotting function called the PlotMainData() to draw the histograms in to the embedded canvas. This function is called inside the ProcessMessage() function in the QwGUIMain class right after the check for the MemoryMap status in the case of M_LOAD_UNLOAD (data load/unload button). In short, once the Load/Unload button is clicked, if the Memory map location is available the histograms will be drawn in the summary tab. To get this to work, I had to activate the SummaryTab inside the QwGUIMain::AddATab() function as soon as it is added to the main tab.

I only tested this on my machine. I Will test it on the cdaq cluster after I implement the error summary plot and make things look "nicer" a bit. 





git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3381 b9700a98-9937-0410-b777-f318d6eb5bff