Cleanup and refactoring of the "qwanalysis.C" script to get a version ready to be run as part of the automatic first100k tools.
To aid development on the farm machines, an overloaded version
of the qwanalysis function which takes the full path of the
ROOT file was added.
A new parameter can be specified to predefine the "hclog_status"
option.

Usages:
	qwanalysis(run_number, <hclog_status>)
	qwanalysis(full_file_name, <run_number>, <hclog_status>)
The hclog_status is optional in both cases; without it, the
script will prompt the user; a value of 1 will submit the
HCLog, and a value of 0 will not submit the HClog.
In the file name version, the run_number is optional, and
defaults to 0; when that happens, the filename is parsed to
determine the run number.
In the file name version, the output files are produced in
the current working directory.



Several functions were added to standardize common operations,
such as generating the framed pads into which the hisotgrams
are drawn.

All of the canvases are drawn by a single HUGE function; a
more scalable solution would be to break each canvas into
its own function which would then be called by the qwanalysis
function.
Further standardization of plotting environments would also
be useful if that were going to be done.

Removed unused variables as reported during ACLiC compilation,
and defined ones which were getting default definitions in CINT.

NOTE ON TESTING:
The output image and text files appear to match what was generated
by the original version, however an exact test will require running
this version and the previous version on the cdaq cluster, where
the previous version is able to find the ROOT files.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3388 b9700a98-9937-0410-b777-f318d6eb5bff