1. Fixed the Bool_t QwBeamLine::Compare(VQwSubsystem *value) routine. It was not checking all the beamline devices.
2. Removed the UpdateErrorCode routine calls to fRelPos, fAbsPos and fEffectiveCharge. No need to update them explicitly since the error codes are ORed when we calculate combined
elements.



3. Following routines were completely commented out for unknown reasons in the QwBPMStripLine class

  void    SetSingleEventCuts(TString ch_name, Double_t minX, Double_t maxX);
  void    SetSingleEventCuts(TString ch_name, UInt_t errorflag,Double_t min, Double_t max, Double_t stability); 

Added the routines properly to pass the event cuts from beamline class into the BPM devices.

4. Added extra check for stability check to see the stability limit is greater than zero.

Summary

Almost finished beamline after fixing some bugs and Halo and Clock devices are left to do which I will do later. 
Next stop is PMT devices. 


Note to myself and others:
Also we don't care about the device error flag at device level. Flags are stored at the channel level. By now the errorflag for a device is not defined uniformly across all the devices and I have not corrected it since we don't care about it anymore. If I get time I could cleaned up this device level error flag. When the error flag is returned to the top level for global event cuts, then the error flags are ORed for all the channels in the device and returned to the top level. 

Rakitha



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/rakithab.new_stability_cuts@3404 b9700a98-9937-0410-b777-f318d6eb5bff